fix: avoid overlap by 0.7*line_width
Depending on which cells are connected to each other, the vertices of the space filling curve should stay away from
the end points of the edge by either line_width/2 or sqrt(2) * line_width/2 .
For safety we are always using sqrt(2) atm.

We could optimize this based on the surrounding triangles and their crossing direction.